requirements.txt
pyrtlsdr>=0.2.91
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
